+++ OpenUtau/Views/MainWindow.axaml.cs
@@ -149,60 +149,7 @@ namespace OpenUtau.App.Views {
             DocManager.Inst.EndUndoGroup();
         }
 
-        /// <summary>
-        /// Remap a tick position from the old time axis to the new time axis without changing its absolute position (in ms).
-        /// Note that this can only be used on positions, not durations.
-        /// </summary>
-        private int RemapTickPos(int tickPos, TimeAxis oldTimeAxis, TimeAxis newTimeAxis){
-            double msPos = oldTimeAxis.TickPosToMsPos(tickPos);
-            return newTimeAxis.MsPosToTickPos(msPos);
-        }
-
-        /// <summary>
-        /// Remap the starting and ending positions of all the notes and parts in the whole project 
-        /// from the old time axis to the new time axis, without changing their absolute positions in ms.
-        /// </summary>
-        private void RemapTimeAxis(TimeAxis oldTimeAxis, TimeAxis newTimeAxis){
-            var project = DocManager.Inst.Project;
-            foreach(var part in project.parts){
-                var partOldStartTick = part.position;
-                var partNewStartTick = RemapTickPos(part.position, oldTimeAxis, newTimeAxis);
-                if(partNewStartTick != partOldStartTick){
-                    DocManager.Inst.ExecuteCmd(new MovePartCommand(
-                        project, part, partNewStartTick, part.trackNo));
-                }
-                if(part is UVoicePart voicePart){
-                    var partOldEndTick = voicePart.End;
-                    var partNewEndTick = RemapTickPos(voicePart.End, oldTimeAxis, newTimeAxis);
-                    if(partNewEndTick - partNewStartTick != voicePart.Duration){
-                        DocManager.Inst.ExecuteCmd(new ResizePartCommand(
-                            project, voicePart, partNewEndTick - partNewStartTick));
-                    }
-                    var noteCommands = new List<UCommand>();
-                    foreach(var note in voicePart.notes){
-                        var noteOldStartTick = note.position + partOldStartTick;
-                        var noteOldEndTick = note.End + partOldStartTick;
-                        var noteOldDuration = note.duration;
-                        var noteNewStartTick = RemapTickPos(noteOldStartTick, oldTimeAxis, newTimeAxis);
-                        var noteNewEndTick = RemapTickPos(noteOldEndTick, oldTimeAxis, newTimeAxis);
-                        var deltaPosTickInPart = (noteNewStartTick - partNewStartTick) - (noteOldStartTick - partOldStartTick);
-                        if(deltaPosTickInPart != 0){
-                            noteCommands.Add(new MoveNoteCommand(voicePart, note, deltaPosTickInPart, 0));
-                        }
-                        var noteNewDuration = noteNewEndTick - noteNewStartTick;
-                        var deltaDur = noteNewDuration - noteOldDuration;
-                        if(deltaDur != 0){
-                            noteCommands.Add(new ResizeNoteCommand(voicePart, note, deltaDur));
-                        }
-                        //TODO: expression curve remapping, control point pitch remapping
-                    }
-                    foreach(var command in noteCommands){
-                        DocManager.Inst.ExecuteCmd(command);
-                    }
-                }
-            }
-        }
-
+        
         
         void OnMenuRemapTimeaxis(object sender, RoutedEventArgs e){
             var project = DocManager.Inst.Project;
@@ -223,7 +170,7 @@ namespace OpenUtau.App.Views {
                             DocManager.Inst.ExecuteCmd(new DelTempoChangeCommand(
                                 project, tempo.position));
                         }
-                        RemapTimeAxis(oldTimeAxis, project.timeAxis.Clone());
+                        viewModel.RemapTimeAxis(oldTimeAxis, project.timeAxis.Clone());
                         DocManager.Inst.EndUndoGroup();
                     }
                 } catch (Exception e) {
